Actualizados

+++ Tema05/Actividad9.java
@@ -3,9 +3,9 @@ package Tema05;
 public class Actividad9 {
     public static void main(String[] args) {
         System.out.print("Introduce un numero: ");
-        long numero  = Long.parseLong(System.console().readLine());
+        int numero  = Integer.parseInt(System.console().readLine());
         
-        long n = numero ;
+        int n = numero ;
         int numeroDeDigitos = 1;
         
         while (n > 10) {
